[DEV-4568] Move tests about release_match module into tests/internal
